+++ components/common/PageLoader/index.tsx
@@ -1,0 +1,13 @@
+import { Loader } from '@mantine/core'
+
+import * as styles from './styles.css'
+
+const PageLoader = () => {
+    return (
+        <div className={styles.loader}>
+            <Loader color='pink' />
+        </div>
+    )
+}
+
+export default PageLoader
