+++ components/common/PageLoader/styles.css.ts
@@ -1,0 +1,8 @@
+import { style } from '@vanilla-extract/css'
+
+export const loader = style({
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    height: '100vh',
+})
